updating js file to IIFE
wrap pokemon list array in an IIFE to avoid accidentally accessing the global state.
create a new pokemonRepository variable to hold what IIFE will return, then assign the IIFE to that variable.
The IIFE return an object with the following public functions assigned as keys:
getAll
add
 functions are defined separately with the function keyword
update the loop code to cope with the new changes.